[IA64] declaration of metaphysical_rr
I'm making a patch for domain destruction.
While I tested my patch by repeatedly 
creating/destructing domains,
a reserved register/field fault occurred in load_region_regs().

The cause of this fault is rr0 value overflow(0xFFFFFFFF********)
because metaphysical_rr0 in the arch_domain struct is 
declared as integer (should be unsigned long).
The attached patch fixes the problem.

+++ xen/include/asm-ia64/domain.h
@@ -15,8 +15,8 @@ extern void domain_relinquish_resources(struct domain *);
 struct arch_domain {
     struct mm_struct *active_mm;
     struct mm_struct *mm;
-    int metaphysical_rr0;
-    int metaphysical_rr4;
+    unsigned long metaphysical_rr0;
+    unsigned long metaphysical_rr4;
     int starting_rid;		/* first RID assigned to domain */
     int ending_rid;		/* one beyond highest RID assigned to domain */
     int rid_bits;		/* number of virtual rid bits (default: 18) */
@@ -61,10 +61,10 @@ struct arch_vcpu {
 	unsigned long xen_timer_interval;
 #endif
     mapped_regs_t *privregs; /* save the state of vcpu */
-    int metaphysical_rr0;		// from arch_domain (so is pinned)
-    int metaphysical_rr4;		// from arch_domain (so is pinned)
-    int metaphysical_saved_rr0;		// from arch_domain (so is pinned)
-    int metaphysical_saved_rr4;		// from arch_domain (so is pinned)
+    unsigned long metaphysical_rr0;		// from arch_domain (so is pinned)
+    unsigned long metaphysical_rr4;		// from arch_domain (so is pinned)
+    unsigned long metaphysical_saved_rr0;	// from arch_domain (so is pinned)
+    unsigned long metaphysical_saved_rr4;	// from arch_domain (so is pinned)
     int breakimm;			// from arch_domain (so is pinned)
     int starting_rid;		/* first RID assigned to domain */
     int ending_rid;		/* one beyond highest RID assigned to domain */
